move swa pool
