+++ python/sglang/srt/mem_cache/memory_pool.py
@@ -40,7 +40,7 @@ KVCache actually holds the physical kv cache.
 import abc
 import logging
 from contextlib import contextmanager, nullcontext
-from typing import TYPE_CHECKING, Dict, List, Optional, Tuple, Union
+from typing import TYPE_CHECKING, List, Optional, Tuple, Union
 
 import numpy as np
 import torch
@@ -1275,209 +1275,6 @@ class HybridLinearKVPool(KVCache):
         assert self.use_mla, "get_mla_kv_buffer called when use_mla is False"
         with self._transfer_id_context(layer):
             return self.full_kv_pool.get_mla_kv_buffer(layer, loc, dst_dtype)
-
-
-class SWAKVPool(KVCache):
-    """KV cache with separate pools for full and SWA attention layers."""
-
-    def __init__(
-        self,
-        size: int,
-        size_swa: int,
-        page_size: int,
-        dtype: torch.dtype,
-        head_num: int,
-        head_dim: int,
-        swa_attention_layer_ids: List[int],
-        full_attention_layer_ids: List[int],
-        enable_kvcache_transpose: bool,
-        device: str,
-        token_to_kv_pool_class: KVCache = MHATokenToKVPool,
-        **kwargs,
-    ):
-        self.size = size
-        self.size_swa = size_swa
-        self.dtype = dtype
-        self.head_num = head_num
-        self.head_dim = head_dim
-        self.device = device
-        self.swa_layer_nums = len(swa_attention_layer_ids)
-        self.full_layer_nums = len(full_attention_layer_ids)
-        self.start_layer = 0
-        self.page_size = page_size
-        self.swa_loc = None
-
-        kwargs["page_size"] = page_size
-        kwargs["enable_memory_saver"] = False
-        kwargs["head_num"] = head_num
-        kwargs["head_dim"] = head_dim
-        kwargs["device"] = device
-        # TODO MHATransposedTokenToKVPool if enable_kvcache_transpose is True
-        assert not enable_kvcache_transpose
-
-        # for disagg with nvlink
-        self.enable_custom_mem_pool, self.custom_mem_pool, _ = (
-            maybe_init_custom_mem_pool(device=self.device)
-        )
-
-        self.swa_kv_pool = token_to_kv_pool_class(
-            size=size_swa,
-            dtype=dtype,
-            layer_num=self.swa_layer_nums,
-            **kwargs,
-        )
-        kwargs.pop("swa_head_num", None)
-        kwargs.pop("swa_head_dim", None)
-        kwargs.pop("swa_v_head_dim", None)
-        self.full_kv_pool = token_to_kv_pool_class(
-            size=size,
-            dtype=dtype,
-            layer_num=self.full_layer_nums,
-            **kwargs,
-        )
-        # {layer_id: (index, is_swa_layer)}
-        self.layers_mapping: Dict[int, Tuple[int, bool]] = {}
-        for full_attn_layer_id, global_layer_id in enumerate(full_attention_layer_ids):
-            self.layers_mapping[global_layer_id] = (full_attn_layer_id, False)
-        for swa_layer_id, global_layer_id in enumerate(swa_attention_layer_ids):
-            self.layers_mapping[global_layer_id] = (swa_layer_id, True)
-        self.full_to_swa_index_mapping: Optional[torch.Tensor] = None
-
-        k_size, v_size = self.get_kv_size_bytes()
-        self.mem_usage = (k_size + v_size) / GB
-        logger.info(
-            f"SWAKVPool mem usage: {self.mem_usage:.2f} GB, swa size: {self.size_swa}, full size: {self.size}"
-        )
-
-    def register_mapping(self, full_to_swa_index_mapping: torch.Tensor):
-        self.full_to_swa_index_mapping = full_to_swa_index_mapping
-
-    def get_kv_size_bytes(self):
-        k_size, v_size = self.full_kv_pool.get_kv_size_bytes()
-        k_size_swa, v_size_swa = self.swa_kv_pool.get_kv_size_bytes()
-        return k_size + k_size_swa, v_size + v_size_swa
-
-    def get_contiguous_buf_infos(self):
-        full_kv_data_ptrs, full_kv_data_lens, full_kv_item_lens = (
-            self.full_kv_pool.get_contiguous_buf_infos()
-        )
-        return (
-            full_kv_data_ptrs,
-            full_kv_data_lens,
-            full_kv_item_lens,
-        )
-
-    def get_state_buf_infos(self):
-        swa_kv_data_ptrs, swa_kv_data_lens, swa_kv_item_lens = (
-            self.swa_kv_pool.get_contiguous_buf_infos()
-        )
-
-        return swa_kv_data_ptrs, swa_kv_data_lens, swa_kv_item_lens
-
-    def get_key_buffer(self, layer_id: int):
-        layer_id_pool, is_swa_layer = self.layers_mapping[layer_id]
-        if is_swa_layer:
-            return self.swa_kv_pool.get_key_buffer(layer_id_pool)
-        else:
-            return self.full_kv_pool.get_key_buffer(layer_id_pool)
-
-    def get_value_buffer(self, layer_id: int):
-        layer_id_pool, is_swa_layer = self.layers_mapping[layer_id]
-        if is_swa_layer:
-            return self.swa_kv_pool.get_value_buffer(layer_id_pool)
-        else:
-            return self.full_kv_pool.get_value_buffer(layer_id_pool)
-
-    def get_kv_buffer(self, layer_id: int):
-        layer_id_pool, is_swa_layer = self.layers_mapping[layer_id]
-        if is_swa_layer:
-            return self.swa_kv_pool.get_kv_buffer(layer_id_pool)
-        else:
-            return self.full_kv_pool.get_kv_buffer(layer_id_pool)
-
-    def set_swa_loc(self, loc: torch.Tensor):
-        self.swa_loc = loc
-
-    def translate_loc_from_full_to_swa(self, kv_indices: torch.Tensor):
-        assert self.full_to_swa_index_mapping is not None
-
-        # Note: kv_indices could have -1 values (from alloc_extend), which will be mapped to -1
-        # since the last item of full_to_swa_index_mapping is -1.
-        return self.full_to_swa_index_mapping[kv_indices].to(torch.int32)
-
-    def set_kv_buffer(
-        self,
-        layer: RadixAttention,
-        loc: torch.Tensor,
-        cache_k: torch.Tensor,
-        cache_v: torch.Tensor,
-        k_scale: float = 1.0,
-        v_scale: float = 1.0,
-    ):
-
-        layer_id = layer.layer_id
-        layer_id_pool, is_swa_layer = self.layers_mapping[layer_id]
-        if is_swa_layer:
-            if self.swa_loc is not None:
-                loc = self.swa_loc
-            else:
-                if self.full_to_swa_index_mapping is not None:
-                    loc = self.translate_loc_from_full_to_swa(loc)
-
-            self.swa_kv_pool.set_kv_buffer(
-                None,
-                loc,
-                cache_k,
-                cache_v,
-                k_scale,
-                v_scale,
-                layer_id_override=layer_id_pool,
-            )
-        else:
-            self.full_kv_pool.set_kv_buffer(
-                None,
-                loc,
-                cache_k,
-                cache_v,
-                k_scale,
-                v_scale,
-                layer_id_override=layer_id_pool,
-            )
-
-    def move_kv_cache(self, tgt_loc: torch.Tensor, src_loc: torch.Tensor):
-        self.full_kv_pool.move_kv_cache(tgt_loc, src_loc)
-        tgt_loc_swa = self.translate_loc_from_full_to_swa(tgt_loc)
-        src_loc_swa = self.translate_loc_from_full_to_swa(src_loc)
-        self.swa_kv_pool.move_kv_cache(tgt_loc_swa, src_loc_swa)
-
-    def get_cpu_copy(self, indices):
-        # For SWA, we need to copy KV cache from both full and SWA pools
-        # The indices are for the full pool, and we use mapping to get SWA indices
-        full_kv_cpu = self.full_kv_pool.get_cpu_copy(indices)
-
-        # Get SWA indices through the mapping
-        # Note: SWA allocation always creates 1:1 mapping, so no need to filter
-        if self.full_to_swa_index_mapping is not None:
-            swa_indices = self.full_to_swa_index_mapping[indices]
-            swa_kv_cpu = self.swa_kv_pool.get_cpu_copy(swa_indices)
-        else:
-            swa_kv_cpu = None
-
-        return {"full": full_kv_cpu, "swa": swa_kv_cpu}
-
-    def load_cpu_copy(self, kv_cache_cpu, indices):
-        # Load KV cache back from CPU to both full and SWA pools
-        # Note: indices here are NEW indices (newly allocated), different from get_cpu_copy indices
-        full_kv_cpu = kv_cache_cpu["full"]
-        swa_kv_cpu = kv_cache_cpu["swa"]
-
-        # Load full KV cache to the new indices
-        self.full_kv_pool.load_cpu_copy(full_kv_cpu, indices)
-
-        # Load SWA KV cache if it exists
-        if swa_kv_cpu is not None and self.full_to_swa_index_mapping is not None:
-            swa_indices = self.full_to_swa_index_mapping[indices]
-            self.swa_kv_pool.load_cpu_copy(swa_kv_cpu, swa_indices)
 
 
 class MLATokenToKVPool(KVCache):
